restructuring for pip package
