+++ wbck/sources/__init__.py
@@ -1,0 +1,10 @@
+from .base import BaseSource
+from .aws import AwsSource
+from .local import LocalSource
+
+
+__all__ = [
+    "BaseSource",
+    "LocalSource",
+    "AwsSource"
+]
